ecommerce/requirements.txt
Django==5.1.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
cryptography==42.0.5
gunicorn==21.2.0
razorpay==1.4.1
